player move&anim upload
