+++ soft_wave_unity/Assets/Animator.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c893df2d9adb51c499abbc1e35828695
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
