+++ soft_wave_unity/Assets/Animator/Player.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37f949bf23419a849bb4dbfe3553b8fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
