+++ soft_wave_unity/Assets/Animator/attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 21c13301bc967d54aa9704fcafd79513
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
