+++ soft_wave_unity/Assets/Animator/damaged.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bdab9f7a63d6ad3489a130f571282ded
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
