+++ soft_wave_unity/Assets/Animator/idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 689c8824732bbc54fb4bc1b0190baa40
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
