+++ soft_wave_unity/Assets/Animator/move.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5557795f461b81044aa355b1cb2813ee
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
