+++ soft_wave_unity/Assets/InputAction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a17aef5932c03cd41b2416c334af6b9a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
