+++ soft_wave_unity/Assets/InputAction/Player.inputactions
@@ -1,0 +1,54 @@
+{
+    "name": "Player",
+    "maps": [
+        {
+            "name": "Move",
+            "id": "fbef8cc1-3e9d-40fe-9f18-781fa1b9b9e7",
+            "actions": [
+                {
+                    "name": "move",
+                    "type": "Button",
+                    "id": "46f4060d-ceb3-48fe-96a7-090cc32ba2d4",
+                    "expectedControlType": "Button",
+                    "processors": "",
+                    "interactions": "",
+                    "initialStateCheck": false
+                },
+                {
+                    "name": "pointer",
+                    "type": "Value",
+                    "id": "0c845668-2d26-4de2-b5e2-9ab90b3f4e1d",
+                    "expectedControlType": "Vector2",
+                    "processors": "",
+                    "interactions": "",
+                    "initialStateCheck": true
+                }
+            ],
+            "bindings": [
+                {
+                    "name": "",
+                    "id": "8ec4ff5d-282c-40fb-899e-b97487dc78b8",
+                    "path": "<Mouse>/rightButton",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "move",
+                    "isComposite": false,
+                    "isPartOfComposite": false
+                },
+                {
+                    "name": "",
+                    "id": "629e5159-512a-4bcf-b7c6-a5ea70155203",
+                    "path": "<Mouse>/position",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "pointer",
+                    "isComposite": false,
+                    "isPartOfComposite": false
+                }
+            ]
+        }
+    ],
+    "controlSchemes": []
+}
