+++ soft_wave_unity/Assets/InputAction/Player.inputactions.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: 6d1096c7adb9515469d64f28ae6b8bba
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
+  generateWrapperCode: 0
+  wrapperCodePath: 
+  wrapperClassName: 
+  wrapperCodeNamespace:
