+++ soft_wave_unity/Assets/Scripts/PlayerAnim.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3cb32c0ef4f9b7b488f142f147d370a2
+guid: 1ee1ab8fd980e90428299cec3901500a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
