+++ soft_wave_unity/Assets/Scripts/PlayerMove.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 05a931086a84b034eb784de2dd0b7db3
+guid: abaa669b97aed544294dc2a064d6ad47
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
